--- frontend/src/components/carFilter3/styles.css
@@ -1,182 +1,0 @@
-.card-section{
-    width: 100%;
-    position: absolute;
-    top: 250px;
-    border: none;
-    z-index: 2;
-    display: flex;
-    justify-content: center;
-}
-.column-card{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    align-content: center;
-    align-self: center;
-    width: 80%;
-    margin: 0 auto;
-}
-
-.mb-3{
-    width: 80%;
-    max-height: 2730px;
-    border: none;
-    border-radius: 20px;
-    -webkit-box-shadow: -1px 7px 29px -8px rgba(0,0,0,0.75);
-    -moz-box-shadow: -1px 7px 29px -8px rgba(0,0,0,0.75);
-    box-shadow: -1px 7px 29px -8px rgba(0,0,0,0.75);
-}
-
-.title-header{
-    font-size: 32px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.btn.btn-danger{
-    margin: 0;
-}
-
-.btn.btn-secondary{
-    margin: 0;
-}
-
-.modal-footer{
-    display: flex;
-    justify-content: space-between;
-}
-
-.card-options{
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-    padding: 5%;
-}
-
-.form-group{
-    width: 100%;
-}
-
-.find-csv{
-    width: 140px;
-    height: 40px;
-}
-
-.base-select{
-    width: 100px;
-    height: 40px;
-}
-
-
-.test-inputo{
-    width: 40%;
-}
-
-.input-block{
-    width: 50%;
-}
-
-.custom-file-label {
-    top: 65px;
-    width: 50%;
-    margin: 0 auto;
-
-}
-
-label{
-    margin: 15px;
-}
-
-.form-control{
-    width: 60%;
-    margin: 0 auto;
-}
-
-.row-selects{
-    display: flex;
-    flex-direction: row;
-}
-
-.btn{
-    margin: 0 auto;
-    margin-top: 30px;
-}
-
-.filter-table{
-    max-height: 2730px;
-}
-
-.filterable-table-container .header-row {
-    margin-top: 6px;
-  }
-  .filterable-table-container .header-row .exact-filters {
-    margin-top: 20px;
-  }
-  .filterable-table-container .header-row .record-count {
-    padding-top: 7px;
-  }
-  .filterable-table-container table.filterable-table thead tr th {
-    white-space: nowrap;
-    user-select: none;
-  }
-  .filterable-table-container table.filterable-table thead tr th span.fa {
-    padding-left: 8px;
-    color: #aaa;
-  }
-  .filterable-table-container table.filterable-table tfoot tr td {
-    font-weight: 700;
-  }
-  .filterable-table-container .filter-container {
-    position: relative;
-  }
-  .filterable-table-container .filter-container .filter-input,
-  .filterable-table-container .filter-container .page-size {
-    width: 185px;
-    display: inline;
-  }
-  .filterable-table-container .clear-filter {
-    position: absolute;
-    right: 5px;
-    top: 1px;
-  }
-  .filterable-table-container span.sortable,
-  .filterable-table-container span.filterable {
-    cursor: pointer;
-  }
-  .filterable-table-container span.filterable {
-    border-bottom: 1px solid #91c1ff;
-  }
-  .filterable-table-container ul.pagination li a {
-    user-select: none;
-  }
-  .filterable-table-container ul.pagination li:not(.disabled) a {
-    cursor: pointer;
-  }
-  .filterable-table-container span.empty {
-    font-style: italic;
-    color: #999;
-  }
-  .filterable-table-container .filter-item {
-    border: 1px solid #ccc;
-    border-radius: 1px;
-    display: inline-block;
-    margin-right: 15px;
-  }
-  .filterable-table-container .filter-item .filter-item-title {
-    background-color: #2f6e9b;
-    border-radius: 1px;
-    color: #fff;
-  }
-  .filterable-table-container .filter-item .filter-item-title .filter-item-remove {
-    color: #fff;
-    font-weight: 700;
-    user-select: none;
-    padding-right: 10px;
-    cursor: pointer;
-  }
-  .filterable-table-container .filter-item .filter-item-title,
-  .filterable-table-container .filter-item .filter-item-value {
-    padding: 5px 10px;
-    display: inline-block;
-  }
