+++ frontend/src/components/cardUpload/index.js
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import './styles.css';
 import Api from '../../services/uploadApi';
-import { Button, Form, FormLabel, FormControl } from 'react-bootstrap';
+import { Button, Form, FormLabel, FormControl, Modal } from 'react-bootstrap';
 import FormFileInput from 'react-bootstrap/FormFileInput';
 import FormFileLabel from 'react-bootstrap/FormFileLabel';
 
@@ -10,17 +10,20 @@ import FormFileLabel from 'react-bootstrap/FormFileLabel';
 const UploadCard = () => {
     const [fileUpload, setFileUpload] = useState([]);
     const [fileName, setFileName] = useState('Procure sua Base');
+    const [show, setShow] = useState(false);
+
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
 
     function onChange(e){
         setFileUpload(e.target.files[0]);
         setFileName(e.target.files[0].name);
-    }
+    };
 
     async function onSubmit(e){
         e.preventDefault();
         const formData = new FormData;
         formData.append('file', fileUpload);
-
         try{
             await Api.uploadApi.post('/base-csv', formData, {
                 headers: {
@@ -28,15 +31,12 @@ const UploadCard = () => {
                 }
             });
             console.log('Arquivo enviado')
-
             setFileUpload([]);
             setFileName('Procure sua Base');
-
         } catch(err){
             console.log(err);
         }
     }
-
 
     return (
         <section className="card-section">
@@ -52,15 +52,25 @@ const UploadCard = () => {
                                 <FormControl as="select">
                                     <option>Webmotors</option>
                                 </FormControl>
-                                <Button type="submit"onSubmit={onSubmit}>Enviar</Button>
+                                <Button type="submit"onSubmit={onSubmit} onClick={handleShow}>Enviar</Button>
+                                    <Modal show={show} onHide={handleClose}>
+                                    <Modal.Header closeButton>
+                                        <Modal.Title>Dados Enviados!</Modal.Title>
+                                    </Modal.Header>
+                                    <Modal.Body>Seu arquivo foi enviado para o banco de dados e em breve estará disponível para consulta</Modal.Body>
+                                    <Modal.Footer>
+                                        <Button variant="primary" onClick={handleClose}>
+                                            Ok!
+                                        </Button>
+                                    </Modal.Footer>
+                                </Modal>
                             </Form>
                         </div>
-                        
                     </div>
                 </div>
             </div>
         </section>
     );
-}
+};
 
 export default UploadCard;
